backend/requirements.txt
# Core Framework
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
duckdb>=0.9.0
thefuzz>=0.20.0
python-Levenshtein>=0.23.0

# AI/ML
openai>=1.3.0
xgboost>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
shap>=0.44.0
lightgbm>=4.0.0

# Visualization
plotly>=5.17.0
kaleido>=0.2.1

# Document Export
python-docx>=1.1.0
fpdf2>=2.7.0
reportlab>=4.0.0
Pillow>=10.0.0

# Web Scraping & External Data
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
wikipedia-api>=0.6.0
feedparser>=6.0.10

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
tiktoken>=0.5.0

# Development
pytest>=7.4.0
fastapi  
uvicorn 
